split mysql visitor
