+++ shardingsphere-sql-parser/shardingsphere-sql-parser-engine/src/main/java/org/apache/shardingsphere/sql/parser/core/extractor/util/RuleName.java
@@ -171,7 +171,77 @@ public enum RuleName {
     
     UPDATE("Update"),
     
-    DELETE("Delete");
+    DELETE("Delete"),
+    
+    REPLACE("Replace"),
+    
+    CREATE_TABLE("CreateTable"),
+    
+    ALTER_TABLE("AlterTable"),
+    
+    DROP_TABLE("DropTable"),
+    
+    TRUNCATE_TABLE("TruncateTable"),
+    
+    CREATE_INDEX("CreateIndex"),
+    
+    DROP_INDEX("DropIndex"),
+    
+    SET_TRANSACTION("SetTransaction"),
+    
+    BEGIN_TRANSACTION("BeginTransaction"),
+    
+    SET_AUTOCOMMIT("SetAutoCommit"),
+    
+    COMMIT("Commit"),
+    
+    ROLLBACK("Rollback"),
+    
+    SAVE_POINT("Savepoint"),
+    
+    GRANT("Grant"),
+    
+    REVOKE("Revoke"),
+    
+    CREATE_USER("CreateUser"),
+    
+    DROP_USER("DropUser"),
+    
+    ALTER_USER("AlterUser"),
+    
+    RENAME_USER("RenameUser"),
+    
+    CREATE_ROLE("CreateRole"),
+    
+    DROP_ROLE("DropRole"),
+    
+    SET_DEFAULT_ROLE("SetDefaultRole"),
+    
+    SET_ROLE("SetRole"),
+    
+    SET_PASSWORD("SetPassword"),
+    
+    USE("Use"),
+    
+    DESC("Desc"),
+    
+    SHOW_DATABASES("ShowDatabases"),
+    
+    SHOW_TABLES("ShowTables"),
+    
+    SHOW_TABLE_STATUS("ShowTableStatus"),
+    
+    SHOW_COLUMNS("ShowColumns"),
+    
+    SHOW_INDEX("ShowIndex"),
+    
+    SHOW_CREATE_TABLE("ShowCreateTable"),
+    
+    SHOW_OTHER("ShowOther"),
+    
+    SET_VARIABLE("SetVariable"),
+    
+    CALL("Call");
     
     private final String name;
