+++ shardingsphere-sql-parser/shardingsphere-sql-parser-mysql/src/main/java/org/apache/shardingsphere/sql/parser/visitor/MySQLTCLVisitor.java
@@ -1,0 +1,86 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.shardingsphere.sql.parser.visitor;
+
+import org.apache.shardingsphere.sql.parser.MySQLVisitor;
+import org.apache.shardingsphere.sql.parser.autogen.MySQLStatementParser.AutoCommitValueContext;
+import org.apache.shardingsphere.sql.parser.autogen.MySQLStatementParser.BeginTransactionContext;
+import org.apache.shardingsphere.sql.parser.autogen.MySQLStatementParser.CommitContext;
+import org.apache.shardingsphere.sql.parser.autogen.MySQLStatementParser.RollbackContext;
+import org.apache.shardingsphere.sql.parser.autogen.MySQLStatementParser.SavepointContext;
+import org.apache.shardingsphere.sql.parser.autogen.MySQLStatementParser.SetAutoCommitContext;
+import org.apache.shardingsphere.sql.parser.autogen.MySQLStatementParser.SetTransactionContext;
+import org.apache.shardingsphere.sql.parser.sql.ASTNode;
+import org.apache.shardingsphere.sql.parser.sql.segment.tcl.AutoCommitSegment;
+import org.apache.shardingsphere.sql.parser.sql.statement.tcl.BeginTransactionStatement;
+import org.apache.shardingsphere.sql.parser.sql.statement.tcl.CommitStatement;
+import org.apache.shardingsphere.sql.parser.sql.statement.tcl.RollbackStatement;
+import org.apache.shardingsphere.sql.parser.sql.statement.tcl.SavepointStatement;
+import org.apache.shardingsphere.sql.parser.sql.statement.tcl.SetAutoCommitStatement;
+import org.apache.shardingsphere.sql.parser.sql.statement.tcl.SetTransactionStatement;
+
+/**
+ * MySQL TCL visitor.
+ *
+ * @author panjuan
+ */
+public final class MySQLTCLVisitor extends MySQLVisitor {
+    
+    @Override
+    public ASTNode visitSetTransaction(final SetTransactionContext ctx) {
+        return new SetTransactionStatement();
+    }
+    
+    @Override
+    public ASTNode visitSetAutoCommit(final SetAutoCommitContext ctx) {
+        SetAutoCommitStatement result = new SetAutoCommitStatement();
+        AutoCommitValueContext autoCommitValueContext = ctx.autoCommitValue();
+        if (null != autoCommitValueContext) {
+            AutoCommitSegment autoCommitSegment = (AutoCommitSegment) visit(ctx.autoCommitValue());
+            result.getAllSQLSegments().add(autoCommitSegment);
+            result.setAutoCommit(autoCommitSegment.isAutoCommit());
+        }
+        return result;
+    }
+    
+    @Override
+    public ASTNode visitAutoCommitValue(final AutoCommitValueContext ctx) {
+        boolean autoCommit = "1".equals(ctx.getText()) || "ON".equals(ctx.getText());
+        return new AutoCommitSegment(ctx.start.getStartIndex(), ctx.stop.getStopIndex(), autoCommit);
+    }
+    
+    @Override
+    public ASTNode visitBeginTransaction(final BeginTransactionContext ctx) {
+        return new BeginTransactionStatement();
+    }
+    
+    @Override
+    public ASTNode visitCommit(final CommitContext ctx) {
+        return new CommitStatement();
+    }
+    
+    @Override
+    public ASTNode visitRollback(final RollbackContext ctx) {
+        return new RollbackStatement();
+    }
+    
+    @Override
+    public ASTNode visitSavepoint(final SavepointContext ctx) {
+        return new SavepointStatement();
+    }
+}
